Features: Implement Get Pet API

+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/schedule/entity/key/IdSchedule.java
@@ -7,8 +7,8 @@ import java.io.Serializable;
 @Embeddable
 public class IdSchedule implements Serializable {
     @Column(name = "pet_id")
-    private Integer petId;
+    private Long petId;
 
     @Column(name = "owner_id")
-    private Integer ownerId;
+    private Long ownerId;
 }
